Modify the 'codelet' prototype in the 'Task'.

+++ include/Module/CRC/CRC.hpp
@@ -6,6 +6,7 @@
 #define CRC_HPP_
 
 #include <cstdint>
+#include <cstddef>
 #include <memory>
 #include <vector>
 
@@ -19,13 +20,13 @@ namespace module
 {
 	namespace crc
 	{
-		enum class tsk : uint8_t { build, extract, check, SIZE };
+		enum class tsk : size_t { build, extract, check, SIZE };
 
 		namespace sck
 		{
-			enum class build   : uint8_t { U_K1, U_K2, SIZE };
-			enum class extract : uint8_t { V_K1, V_K2, SIZE };
-			enum class check   : uint8_t { V_K       , SIZE };
+			enum class build   : size_t { U_K1, U_K2, SIZE };
+			enum class extract : size_t { V_K1, V_K2, SIZE };
+			enum class check   : size_t { V_K       , SIZE };
 		}
 	}
