+++ include/Module/Puncturer/Puncturer.hpp
@@ -19,12 +19,12 @@ namespace module
 {
 	namespace pct
 	{
-		enum class tsk : uint8_t { puncture, depuncture, SIZE };
+		enum class tsk : size_t { puncture, depuncture, SIZE };
 
 		namespace sck
 		{
-			enum class puncture   : uint8_t { X_N1, X_N2, SIZE };
-			enum class depuncture : uint8_t { Y_N1, Y_N2, SIZE };
+			enum class puncture   : size_t { X_N1, X_N2, SIZE };
+			enum class depuncture : size_t { Y_N1, Y_N2, SIZE };
 		}
 	}
